+++ index.js
@@ -1,76 +1,71 @@
-const $ = { enabled:true };
-
 const CODES = {
-  // modifiers
-  reset: fmt(0, 0),
-  bold: fmt(1, 22),
-  dim: fmt(2, 22),
-  italic: fmt(3, 23),
-  underline: fmt(4, 24),
-  inverse: fmt(7, 27),
-  hidden: fmt(8, 28),
-  strikethrough: fmt(9, 29),
-  // colors
-  black: fmt(30, 39),
-  red: fmt(31, 39),
-  green: fmt(32, 39),
-  yellow: fmt(33, 39),
-  blue: fmt(34, 39),
-  magenta: fmt(35, 39),
-  cyan: fmt(36, 39),
-  white: fmt(37, 39),
-  gray: fmt(90, 39),
-  // background colors
-  bgBlack: fmt(40, 49),
-  bgRed: fmt(41, 49),
-  bgGreen: fmt(42, 49),
-  bgYellow: fmt(43, 49),
-  bgBlue: fmt(44, 49),
-  bgMagenta: fmt(45, 49),
-  bgCyan: fmt(46, 49),
-  bgWhite: fmt(47, 49)
+	// modifiers
+	reset: [0, 0],
+	bold: [1, 22],
+	dim: [2, 22],
+	italic: [3, 23],
+	underline: [4, 24],
+	inverse: [7, 27],
+	hidden: [8, 28],
+	strikethrough: [9, 29],
+
+	// colors
+	black: [30, 39],
+	red: [31, 39],
+	green: [32, 39],
+	yellow: [33, 39],
+	blue: [34, 39],
+	magenta: [35, 39],
+	cyan: [36, 39],
+	white: [37, 39],
+	gray: [90, 39],
+
+	// background colors
+	bgBlack: [40, 49],
+	bgRed: [41, 49],
+	bgGreen: [42, 49],
+	bgYellow: [43, 49],
+	bgBlue: [44, 49],
+	bgMagenta: [45, 49],
+	bgCyan: [46, 49],
+	bgWhite: [47, 49]
 };
 
-function fmt(x, y) {
-	return {
-		open: `\x1b[${x}m`,
-		close: `\x1b[${y}m`,
-		rgx: new RegExp(`\\x1b\\[${y}m`, 'g')
-	}
-}
+let $ = { enabled:true };
 
-function run(key, str) {
-	let tmp = CODES[key];
-	return tmp.open + str.replace(tmp.rgx, tmp.open) + tmp.close;
-}
-
-function exec(key, str) {
-	str += '';
-	if (!$.enabled) return str;
-	let arr = this.keys;
+function run(ctx, str) {
+	let tmp;
+	let arr = ctx.keys;
 	while (arr.length > 0) {
-		str = run(arr.shift(), str);
+		tmp = CODES[ arr.shift() ];
+		str = tmp.open + str.replace(tmp.rgx, tmp.open) + tmp.close;
 	}
-	this.keys.push(key);
 	return str;
 }
 
-function attach(key) {
-	let ctx = { keys:[key] };
-	let fn = exec.bind(ctx, key);
-	for (let k in CODES) {
-		Reflect.defineProperty(fn, k, {
-			get() {
-				ctx.keys.push(k);
-				return fn;
+function apply(key, txt) {
+	let str = txt == null ? '' : txt+'';
+	if (!$.enabled) return str || this;
+	if (this.keys === void 0) {
+		this.keys = [];
+		for (let k in $) {
+			if (k !== 'enabled') {
+				this[k] = apply.bind(this, k);
 			}
-		});
+		}
 	}
-	return fn;
+	this.keys.includes(key) || this.keys.push(key);
+	return str ? run(this, str) : this;
 }
 
-for (let k in CODES) {
-	$[k] = attach(k);
+for (let key in CODES) {
+	CODES[key] = {
+		open: `\x1b[${CODES[key][0]}m`,
+		close: `\x1b[${CODES[key][1]}m`,
+		rgx: new RegExp(`\\x1b\\[${CODES[key][1]}m`, 'g')
+	};
+
+	$[key] = apply.bind({}, key);
 }
 
 module.exports = $;
